quic: separate stats and symbols into separate files
