+++ lib/internal/quic/quic.js
@@ -8,21 +8,18 @@ const {
   ArrayBuffer,
   ArrayIsArray,
   ArrayPrototypePush,
-  BigUint64Array,
   DataView,
   DataViewPrototypeGetBigInt64,
   DataViewPrototypeGetBigUint64,
   DataViewPrototypeGetUint8,
   DataViewPrototypeSetUint8,
   ObjectDefineProperties,
-  Symbol,
   SymbolAsyncDispose,
   Uint8Array,
 } = primordials;
 
 const {
   assertCrypto,
-  customInspectSymbol: kInspect,
 } = require('internal/util');
 const { inspect } = require('internal/util/inspect');
 assertCrypto();
@@ -54,62 +51,6 @@ const {
   CLOSECONTEXT_RECEIVE_FAILURE: kCloseContextReceiveFailure,
   CLOSECONTEXT_SEND_FAILURE: kCloseContextSendFailure,
   CLOSECONTEXT_START_FAILURE: kCloseContextStartFailure,
-
-  // All of the IDX_STATS_* constants are the index positions of the stats
-  // fields in the relevant BigUint64Array's that underlie the *Stats objects.
-  // These are not exposed to end users.
-  IDX_STATS_ENDPOINT_CREATED_AT,
-  IDX_STATS_ENDPOINT_DESTROYED_AT,
-  IDX_STATS_ENDPOINT_BYTES_RECEIVED,
-  IDX_STATS_ENDPOINT_BYTES_SENT,
-  IDX_STATS_ENDPOINT_PACKETS_RECEIVED,
-  IDX_STATS_ENDPOINT_PACKETS_SENT,
-  IDX_STATS_ENDPOINT_SERVER_SESSIONS,
-  IDX_STATS_ENDPOINT_CLIENT_SESSIONS,
-  IDX_STATS_ENDPOINT_SERVER_BUSY_COUNT,
-  IDX_STATS_ENDPOINT_RETRY_COUNT,
-  IDX_STATS_ENDPOINT_VERSION_NEGOTIATION_COUNT,
-  IDX_STATS_ENDPOINT_STATELESS_RESET_COUNT,
-  IDX_STATS_ENDPOINT_IMMEDIATE_CLOSE_COUNT,
-
-  IDX_STATS_SESSION_CREATED_AT,
-  IDX_STATS_SESSION_CLOSING_AT,
-  IDX_STATS_SESSION_DESTROYED_AT,
-  IDX_STATS_SESSION_HANDSHAKE_COMPLETED_AT,
-  IDX_STATS_SESSION_HANDSHAKE_CONFIRMED_AT,
-  IDX_STATS_SESSION_GRACEFUL_CLOSING_AT,
-  IDX_STATS_SESSION_BYTES_RECEIVED,
-  IDX_STATS_SESSION_BYTES_SENT,
-  IDX_STATS_SESSION_BIDI_IN_STREAM_COUNT,
-  IDX_STATS_SESSION_BIDI_OUT_STREAM_COUNT,
-  IDX_STATS_SESSION_UNI_IN_STREAM_COUNT,
-  IDX_STATS_SESSION_UNI_OUT_STREAM_COUNT,
-  IDX_STATS_SESSION_LOSS_RETRANSMIT_COUNT,
-  IDX_STATS_SESSION_MAX_BYTES_IN_FLIGHT,
-  IDX_STATS_SESSION_BYTES_IN_FLIGHT,
-  IDX_STATS_SESSION_BLOCK_COUNT,
-  IDX_STATS_SESSION_CWND,
-  IDX_STATS_SESSION_LATEST_RTT,
-  IDX_STATS_SESSION_MIN_RTT,
-  IDX_STATS_SESSION_RTTVAR,
-  IDX_STATS_SESSION_SMOOTHED_RTT,
-  IDX_STATS_SESSION_SSTHRESH,
-  IDX_STATS_SESSION_DATAGRAMS_RECEIVED,
-  IDX_STATS_SESSION_DATAGRAMS_SENT,
-  IDX_STATS_SESSION_DATAGRAMS_ACKNOWLEDGED,
-  IDX_STATS_SESSION_DATAGRAMS_LOST,
-
-  IDX_STATS_STREAM_CREATED_AT,
-  IDX_STATS_STREAM_RECEIVED_AT,
-  IDX_STATS_STREAM_ACKED_AT,
-  IDX_STATS_STREAM_CLOSING_AT,
-  IDX_STATS_STREAM_DESTROYED_AT,
-  IDX_STATS_STREAM_BYTES_RECEIVED,
-  IDX_STATS_STREAM_BYTES_SENT,
-  IDX_STATS_STREAM_MAX_OFFSET,
-  IDX_STATS_STREAM_MAX_OFFSET_ACK,
-  IDX_STATS_STREAM_MAX_OFFSET_RECV,
-  IDX_STATS_STREAM_FINAL_SIZE,
 
   IDX_STATE_SESSION_PATH_VALIDATION,
   IDX_STATE_SESSION_VERSION_NEGOTIATION,
@@ -181,11 +122,6 @@ const {
 } = require('internal/crypto/keys');
 
 const {
-  kHandle: kKeyObjectHandle,
-  kKeyObject: kKeyObjectInner,
-} = require('internal/crypto/util');
-
-const {
   validateFunction,
   validateObject,
 } = require('internal/validators');
@@ -193,21 +129,33 @@ const {
 const kEmptyObject = { __proto__: null };
 const kEnumerable = { __proto__: null, enumerable: true };
 
-const kBlocked = Symbol('kBlocked');
-const kDatagram = Symbol('kDatagram');
-const kDatagramStatus = Symbol('kDatagramStatus');
-const kError = Symbol('kError');
-const kFinishClose = Symbol('kFinishClose');
-const kHandshake = Symbol('kHandshake');
-const kHeaders = Symbol('kHeaders');
-const kOwner = Symbol('kOwner');
-const kNewSession = Symbol('kNewSession');
-const kNewStream = Symbol('kNewStream');
-const kPathValidation = Symbol('kPathValidation');
-const kReset = Symbol('kReset');
-const kSessionTicket = Symbol('kSessionTicket');
-const kTrailers = Symbol('kTrailers');
-const kVersionNegotiation = Symbol('kVersionNegotiation');
+const {
+  kBlocked,
+  kDatagram,
+  kDatagramStatus,
+  kError,
+  kFinishClose,
+  kHandshake,
+  kHeaders,
+  kOwner,
+  kNewSession,
+  kNewStream,
+  kPathValidation,
+  kReset,
+  kSessionTicket,
+  kTrailers,
+  kVersionNegotiation,
+  kInspect,
+  kKeyObjectHandle,
+  kKeyObjectInner,
+  kPrivateConstructor,
+} = require('internal/quic/symbols');
+
+const {
+  EndpointStats,
+  QuicStreamStats,
+  SessionStats,
+} = require('internal/quic/stats');
 
 /**
  * @typedef {import('../socketaddress.js').SocketAddress} SocketAddress
@@ -574,123 +522,6 @@ function processTlsOptions(tls) {
   };
 }
 
-class QuicStreamStats {
-  /** @type {BigUint64Array} */
-  #handle;
-
-  /**
-   * @param {ArrayBuffer} buffer
-   */
-  constructor(buffer) {
-    if (!isArrayBuffer(buffer)) {
-      throw new ERR_INVALID_ARG_TYPE('buffer', ['ArrayBuffer'], buffer);
-    }
-    this.#handle = new BigUint64Array(buffer);
-  }
-
-  /** @type {bigint} */
-  get createdAt() {
-    return this.#handle[IDX_STATS_STREAM_CREATED_AT];
-  }
-
-  /** @type {bigint} */
-  get receivedAt() {
-    return this.#handle[IDX_STATS_STREAM_RECEIVED_AT];
-  }
-
-  /** @type {bigint} */
-  get ackedAt() {
-    return this.#handle[IDX_STATS_STREAM_ACKED_AT];
-  }
-
-  /** @type {bigint} */
-  get closingAt() {
-    return this.#handle[IDX_STATS_STREAM_CLOSING_AT];
-  }
-
-  /** @type {bigint} */
-  get destroyedAt() {
-    return this.#handle[IDX_STATS_STREAM_DESTROYED_AT];
-  }
-
-  /** @type {bigint} */
-  get bytesReceived() {
-    return this.#handle[IDX_STATS_STREAM_BYTES_RECEIVED];
-  }
-
-  /** @type {bigint} */
-  get bytesSent() {
-    return this.#handle[IDX_STATS_STREAM_BYTES_SENT];
-  }
-
-  /** @type {bigint} */
-  get maxOffset() {
-    return this.#handle[IDX_STATS_STREAM_MAX_OFFSET];
-  }
-
-  /** @type {bigint} */
-  get maxOffsetAcknowledged() {
-    return this.#handle[IDX_STATS_STREAM_MAX_OFFSET_ACK];
-  }
-
-  /** @type {bigint} */
-  get maxOffsetReceived() {
-    return this.#handle[IDX_STATS_STREAM_MAX_OFFSET_RECV];
-  }
-
-  /** @type {bigint} */
-  get finalSize() {
-    return this.#handle[IDX_STATS_STREAM_FINAL_SIZE];
-  }
-
-  toJSON() {
-    return {
-      __proto__: null,
-      createdAt: `${this.createdAt}`,
-      receivedAt: `${this.receivedAt}`,
-      ackedAt: `${this.ackedAt}`,
-      closingAt: `${this.closingAt}`,
-      destroyedAt: `${this.destroyedAt}`,
-      bytesReceived: `${this.bytesReceived}`,
-      bytesSent: `${this.bytesSent}`,
-      maxOffset: `${this.maxOffset}`,
-      maxOffsetAcknowledged: `${this.maxOffsetAcknowledged}`,
-      maxOffsetReceived: `${this.maxOffsetReceived}`,
-      finalSize: `${this.finalSize}`,
-    };
-  }
-
-  [kInspect](depth, options) {
-    if (depth < 0)
-      return this;
-
-    const opts = {
-      ...options,
-      depth: options.depth == null ? null : options.depth - 1,
-    };
-
-    return `StreamStats ${inspect({
-      createdAt: this.createdAt,
-      receivedAt: this.receivedAt,
-      ackedAt: this.ackedAt,
-      closingAt: this.closingAt,
-      destroyedAt: this.destroyedAt,
-      bytesReceived: this.bytesReceived,
-      bytesSent: this.bytesSent,
-      maxOffset: this.maxOffset,
-      maxOffsetAcknowledged: this.maxOffsetAcknowledged,
-      maxOffsetReceived: this.maxOffsetReceived,
-      finalSize: this.finalSize,
-    }, opts)}`;
-  }
-
-  [kFinishClose]() {
-    // Snapshot the stats into a new BigUint64Array since the underlying
-    // buffer will be destroyed.
-    this.#handle = new BigUint64Array(this.#handle);
-  }
-}
-
 class QuicStreamState {
   /** @type {DataView} */
   #handle;
@@ -870,7 +701,7 @@ class QuicStream {
    */
   constructor(config, handle, session, direction) {
     validateObject(config, 'config');
-    this.#stats = new QuicStreamStats(handle.stats);
+    this.#stats = new QuicStreamStats(kPrivateConstructor, handle.stats);
     this.#state = new QuicStreamState(handle.stats);
     const {
       onblocked,
@@ -960,228 +791,6 @@ class QuicStream {
       state: this.state,
       session: this.session,
     }, opts)}`;
-  }
-}
-
-class SessionStats {
-  /** @type {BigUint64Array} */
-  #handle;
-
-  /**
-   * @param {BigUint64Array} buffer
-   */
-  constructor(buffer) {
-    if (!isArrayBuffer(buffer)) {
-      throw new ERR_INVALID_ARG_TYPE('buffer', ['ArrayBuffer'], buffer);
-    }
-    this.#handle = new BigUint64Array(buffer);
-  }
-
-  /** @type {bigint} */
-  get createdAt() {
-    return this.#handle[IDX_STATS_SESSION_CREATED_AT];
-  }
-
-  /** @type {bigint} */
-  get closingAt() {
-    return this.#handle[IDX_STATS_SESSION_CLOSING_AT];
-  }
-
-  /** @type {bigint} */
-  get destroyedAt() {
-    return this.#handle[IDX_STATS_SESSION_DESTROYED_AT];
-  }
-
-  /** @type {bigint} */
-  get handshakeCompletedAt() {
-    return this.#handle[IDX_STATS_SESSION_HANDSHAKE_COMPLETED_AT];
-  }
-
-  /** @type {bigint} */
-  get handshakeConfirmedAt() {
-    return this.#handle[IDX_STATS_SESSION_HANDSHAKE_CONFIRMED_AT];
-  }
-
-  /** @type {bigint} */
-  get gracefulClosingAt() {
-    return this.#handle[IDX_STATS_SESSION_GRACEFUL_CLOSING_AT];
-  }
-
-  /** @type {bigint} */
-  get bytesReceived() {
-    return this.#handle[IDX_STATS_SESSION_BYTES_RECEIVED];
-  }
-
-  /** @type {bigint} */
-  get bytesSent() {
-    return this.#handle[IDX_STATS_SESSION_BYTES_SENT];
-  }
-
-  /** @type {bigint} */
-  get bidiInStreamCount() {
-    return this.#handle[IDX_STATS_SESSION_BIDI_IN_STREAM_COUNT];
-  }
-
-  /** @type {bigint} */
-  get bidiOutStreamCount() {
-    return this.#handle[IDX_STATS_SESSION_BIDI_OUT_STREAM_COUNT];
-  }
-
-  /** @type {bigint} */
-  get uniInStreamCount() {
-    return this.#handle[IDX_STATS_SESSION_UNI_IN_STREAM_COUNT];
-  }
-
-  /** @type {bigint} */
-  get uniOutStreamCount() {
-    return this.#handle[IDX_STATS_SESSION_UNI_OUT_STREAM_COUNT];
-  }
-
-  /** @type {bigint} */
-  get lossRetransmitCount() {
-    return this.#handle[IDX_STATS_SESSION_LOSS_RETRANSMIT_COUNT];
-  }
-
-  /** @type {bigint} */
-  get maxBytesInFlights() {
-    return this.#handle[IDX_STATS_SESSION_MAX_BYTES_IN_FLIGHT];
-  }
-
-  /** @type {bigint} */
-  get bytesInFlight() {
-    return this.#handle[IDX_STATS_SESSION_BYTES_IN_FLIGHT];
-  }
-
-  /** @type {bigint} */
-  get blockCount() {
-    return this.#handle[IDX_STATS_SESSION_BLOCK_COUNT];
-  }
-
-  /** @type {bigint} */
-  get cwnd() {
-    return this.#handle[IDX_STATS_SESSION_CWND];
-  }
-
-  /** @type {bigint} */
-  get latestRtt() {
-    return this.#handle[IDX_STATS_SESSION_LATEST_RTT];
-  }
-
-  /** @type {bigint} */
-  get minRtt() {
-    return this.#handle[IDX_STATS_SESSION_MIN_RTT];
-  }
-
-  /** @type {bigint} */
-  get rttVar() {
-    return this.#handle[IDX_STATS_SESSION_RTTVAR];
-  }
-
-  /** @type {bigint} */
-  get smoothedRtt() {
-    return this.#handle[IDX_STATS_SESSION_SMOOTHED_RTT];
-  }
-
-  /** @type {bigint} */
-  get ssthresh() {
-    return this.#handle[IDX_STATS_SESSION_SSTHRESH];
-  }
-
-  /** @type {bigint} */
-  get datagramsReceived() {
-    return this.#handle[IDX_STATS_SESSION_DATAGRAMS_RECEIVED];
-  }
-
-  /** @type {bigint} */
-  get datagramsSent() {
-    return this.#handle[IDX_STATS_SESSION_DATAGRAMS_SENT];
-  }
-
-  /** @type {bigint} */
-  get datagramsAcknowledged() {
-    return this.#handle[IDX_STATS_SESSION_DATAGRAMS_ACKNOWLEDGED];
-  }
-
-  /** @type {bigint} */
-  get datagramsLost() {
-    return this.#handle[IDX_STATS_SESSION_DATAGRAMS_LOST];
-  }
-
-  toJSON() {
-    return {
-      __proto__: null,
-      createdAt: `${this.createdAt}`,
-      closingAt: `${this.closingAt}`,
-      destroyedAt: `${this.destroyedAt}`,
-      handshakeCompletedAt: `${this.handshakeCompletedAt}`,
-      handshakeConfirmedAt: `${this.handshakeConfirmedAt}`,
-      gracefulClosingAt: `${this.gracefulClosingAt}`,
-      bytesReceived: `${this.bytesReceived}`,
-      bytesSent: `${this.bytesSent}`,
-      bidiInStreamCount: `${this.bidiInStreamCount}`,
-      bidiOutStreamCount: `${this.bidiOutStreamCount}`,
-      uniInStreamCount: `${this.uniInStreamCount}`,
-      uniOutStreamCount: `${this.uniOutStreamCount}`,
-      lossRetransmitCount: `${this.lossRetransmitCount}`,
-      maxBytesInFlights: `${this.maxBytesInFlights}`,
-      bytesInFlight: `${this.bytesInFlight}`,
-      blockCount: `${this.blockCount}`,
-      cwnd: `${this.cwnd}`,
-      latestRtt: `${this.latestRtt}`,
-      minRtt: `${this.minRtt}`,
-      rttVar: `${this.rttVar}`,
-      smoothedRtt: `${this.smoothedRtt}`,
-      ssthresh: `${this.ssthresh}`,
-      datagramsReceived: `${this.datagramsReceived}`,
-      datagramsSent: `${this.datagramsSent}`,
-      datagramsAcknowledged: `${this.datagramsAcknowledged}`,
-      datagramsLost: `${this.datagramsLost}`,
-    };
-  }
-
-  [kInspect](depth, options) {
-    if (depth < 0)
-      return this;
-
-    const opts = {
-      ...options,
-      depth: options.depth == null ? null : options.depth - 1,
-    };
-
-    return `SessionStats ${inspect({
-      createdAt: this.createdAt,
-      closingAt: this.closingAt,
-      destroyedAt: this.destroyedAt,
-      handshakeCompletedAt: this.handshakeCompletedAt,
-      handshakeConfirmedAt: this.handshakeConfirmedAt,
-      gracefulClosingAt: this.gracefulClosingAt,
-      bytesReceived: this.bytesReceived,
-      bytesSent: this.bytesSent,
-      bidiInStreamCount: this.bidiInStreamCount,
-      bidiOutStreamCount: this.bidiOutStreamCount,
-      uniInStreamCount: this.uniInStreamCount,
-      uniOutStreamCount: this.uniOutStreamCount,
-      lossRetransmitCount: this.lossRetransmitCount,
-      maxBytesInFlights: this.maxBytesInFlights,
-      bytesInFlight: this.bytesInFlight,
-      blockCount: this.blockCount,
-      cwnd: this.cwnd,
-      latestRtt: this.latestRtt,
-      minRtt: this.minRtt,
-      rttVar: this.rttVar,
-      smoothedRtt: this.smoothedRtt,
-      ssthresh: this.ssthresh,
-      datagramsReceived: this.datagramsReceived,
-      datagramsSent: this.datagramsSent,
-      datagramsAcknowledged: this.datagramsAcknowledged,
-      datagramsLost: this.datagramsLost,
-    }, opts)}`;
-  }
-
-  [kFinishClose]() {
-    // Snapshot the stats into a new BigUint64Array since the underlying
-    // buffer will be destroyed.
-    this.#handle = new BigUint64Array(this.#handle);
   }
 }
 
@@ -1392,7 +1001,7 @@ class Session {
    */
   constructor(config, streamConfig, handle, endpoint) {
     validateObject(config, 'config');
-    this.#stats = new SessionStats(handle.stats);
+    this.#stats = new SessionStats(kPrivateConstructor, handle.stats);
     this.#state = new SessionState(handle.state);
     const {
       ondatagram,
@@ -1688,137 +1297,6 @@ class Session {
   async [SymbolAsyncDispose]() { await this.close(); }
 }
 
-class EndpointStats {
-  /** @type {BigUint64Array} */
-  #handle;
-
-  /**
-   * @param {ArrayBuffer} buffer
-   */
-  constructor(buffer) {
-    if (!isArrayBuffer(buffer)) {
-      throw new ERR_INVALID_ARG_TYPE('buffer', ['ArrayBuffer'], buffer);
-    }
-    this.#handle = new BigUint64Array(buffer);
-  }
-
-  /** @type {bigint} */
-  get createdAt() {
-    return this.#handle[IDX_STATS_ENDPOINT_CREATED_AT];
-  }
-
-  /** @type {bigint} */
-  get destroyedAt() {
-    return this.#handle[IDX_STATS_ENDPOINT_DESTROYED_AT];
-  }
-
-  /** @type {bigint} */
-  get bytesReceived() {
-    return this.#handle[IDX_STATS_ENDPOINT_BYTES_RECEIVED];
-  }
-
-  /** @type {bigint} */
-  get bytesSent() {
-    return this.#handle[IDX_STATS_ENDPOINT_BYTES_SENT];
-  }
-
-  /** @type {bigint} */
-  get packetsReceived() {
-    return this.#handle[IDX_STATS_ENDPOINT_PACKETS_RECEIVED];
-  }
-
-  /** @type {bigint} */
-  get packetsSent() {
-    return this.#handle[IDX_STATS_ENDPOINT_PACKETS_SENT];
-  }
-
-  /** @type {bigint} */
-  get serverSessions() {
-    return this.#handle[IDX_STATS_ENDPOINT_SERVER_SESSIONS];
-  }
-
-  /** @type {bigint} */
-  get clientSessions() {
-    return this.#handle[IDX_STATS_ENDPOINT_CLIENT_SESSIONS];
-  }
-
-  /** @type {bigint} */
-  get serverBusyCount() {
-    return this.#handle[IDX_STATS_ENDPOINT_SERVER_BUSY_COUNT];
-  }
-
-  /** @type {bigint} */
-  get retryCount() {
-    return this.#handle[IDX_STATS_ENDPOINT_RETRY_COUNT];
-  }
-
-  /** @type {bigint} */
-  get versionNegotiationCount() {
-    return this.#handle[IDX_STATS_ENDPOINT_VERSION_NEGOTIATION_COUNT];
-  }
-
-  /** @type {bigint} */
-  get statelessResetCount() {
-    return this.#handle[IDX_STATS_ENDPOINT_STATELESS_RESET_COUNT];
-  }
-
-  /** @type {bigint} */
-  get immediateCloseCount() {
-    return this.#handle[IDX_STATS_ENDPOINT_IMMEDIATE_CLOSE_COUNT];
-  }
-
-  toJSON() {
-    return {
-      __proto__: null,
-      createdAt: `${this.createdAt}`,
-      destroyedAt: `${this.destroyedAt}`,
-      bytesReceived: `${this.bytesReceived}`,
-      bytesSent: `${this.bytesSent}`,
-      packetsReceived: `${this.packetsReceived}`,
-      packetsSent: `${this.packetsSent}`,
-      serverSessions: `${this.serverSessions}`,
-      clientSessions: `${this.clientSessions}`,
-      serverBusyCount: `${this.serverBusyCount}`,
-      retryCount: `${this.retryCount}`,
-      versionNegotiationCount: `${this.versionNegotiationCount}`,
-      statelessResetCount: `${this.statelessResetCount}`,
-      immediateCloseCount: `${this.immediateCloseCount}`,
-    };
-  }
-
-  [kInspect](depth, options) {
-    if (depth < 0)
-      return this;
-
-    const opts = {
-      ...options,
-      depth: options.depth == null ? null : options.depth - 1,
-    };
-
-    return `EndpointStats ${inspect({
-      createdAt: this.createdAt,
-      destroyedAt: this.destroyedAt,
-      bytesReceived: this.bytesReceived,
-      bytesSent: this.bytesSent,
-      packetsReceived: this.packetsReceived,
-      packetsSent: this.packetsSent,
-      serverSessions: this.serverSessions,
-      clientSessions: this.clientSessions,
-      serverBusyCount: this.serverBusyCount,
-      retryCount: this.retryCount,
-      versionNegotiationCount: this.versionNegotiationCount,
-      statelessResetCount: this.statelessResetCount,
-      immediateCloseCount: this.immediateCloseCount,
-    }, opts)}`;
-  }
-
-  [kFinishClose]() {
-    // Snapshot the stats into a new BigUint64Array since the underlying
-    // buffer will be destroyed.
-    this.#handle = new BigUint64Array(this.#handle);
-  }
-}
-
 class EndpointState {
   /** @type {DataView} */
   #handle;
@@ -2074,7 +1552,7 @@ class Endpoint {
       tokenSecret,
     });
     this.#handle[kOwner] = this;
-    this.#stats = new EndpointStats(this.#handle.stats);
+    this.#stats = new EndpointStats(kPrivateConstructor, this.#handle.stats);
     this.#state = new EndpointState(this.#handle.state);
   }
